refactor: mongo version key removed
