+++ src/courses/courses.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { CoursesService } from './courses.service';
-import { CoursesController } from './coursesController';
+import { CoursesController } from './courses.controller';
 import { MongooseModule } from '@nestjs/mongoose';
 import { Course, CourseSchema } from './entities/courses.entity';
