+++ src/mail/mail.controller.ts
@@ -4,6 +4,7 @@ import RoleGuard from '~/authentication/guards/role.guard';
 import { Roles } from '~/common/constants';
 import { SendEmailDTO } from './dto';
 import { MailService } from './mail.service';
+import { ApiExcludeEndpoint } from '@nestjs/swagger';
 
 @Controller('mail')
 export class MailController {
@@ -11,6 +12,7 @@ export class MailController {
 
   @UseGuards(RoleGuard(Roles.Admin))
   @UseGuards(AccessTokenGuard)
+  @ApiExcludeEndpoint()
   @Post('send-email')
   /* TODO: implement multiple email sending */
   async sendEmail(@Body() params: SendEmailDTO) {
