+++ src/majors/entities/major.entity.ts
@@ -3,10 +3,10 @@ import { Document } from 'mongoose';
 
 export type MajorDocument = Major & Document;
 
-@Schema({ timestamps: true })
+@Schema({ timestamps: true, versionKey: false })
 export class Major {
-    @Prop({ required: true })
-    name: string;
+  @Prop({ required: true })
+  name: string;
 }
 
 export const MajorSchema = SchemaFactory.createForClass(Major);
